requirements.txt
fastapi
uvicorn
insightface
opencv-python
numpy
faiss-cpu
pillow
onnxruntime
filelock
python-dotenv
celery
redis
sqlalchemy
psycopg2-binary
boto3
python-multipart